+++ unliftio/ChangeLog.md
@@ -1,3 +1,7 @@
+## 0.2.2.0
+
+* Add `pureTry` and `pureTryDeep`
+
 ## 0.2.1.0
 
 * Add `UnliftIO.STM`